requirements.txt
openai-agents==0.1.0
pydantic>=2.1.0,<3.0
python-dotenv==1.0.1
